# newgram-ui/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Newgram API",
    "description": "API para Newgram, uma aplicação de rede social inspirada no Instagram.",
    "termsOfService": "https://example.com/terms",
    "contact": {
      "name": "Jameson Henrique",
      "url": "https://github.com/JamesonHenrique",
      "email": "jamesonhenrique14@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    },
    "version": "1.0.0"
  },
...
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Servidor de desenvolvimento"
    }
  ],
  "security": [{ "bearerAuth": [] }],
...
  "paths": {
    "/posts/{id}": {
      "get": {
        "tags": ["Posts"],
        "summary": "Buscar post por ID",
        "description": "Retorna os detalhes de um post específico",
        "operationId": "buscarPorId",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID do post",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          }
        ],
        "responses": {
          "200": {
            "description": "Post encontrado com sucesso",
            "content": {
              "*/*": {
                "schema": { "$ref": "#/components/schemas/PostResponseDTO" }
              }
            }
          },
          "401": { "description": "Não autorizado" },
          "404": { "description": "Post não encontrado" }
        }
      },
      "put": {
        "tags": ["Posts"],
        "summary": "Atualizar post",
        "description": "Atualiza um post existente",
        "operationId": "atualizarPost",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID do post",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PostUpdateDTO" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Post atualizado com sucesso",
            "content": {
              "*/*": {
                "schema": { "$ref": "#/components/schemas/PostResponseDTO" }
              }
            }
          },
          "400": { "description": "Dados inválidos fornecidos" },
          "401": { "description": "Não autorizado" },
          "403": { "description": "Permissão negada" },
          "404": { "description": "Post não encontrado" }
        }
      },
      "delete": {
        "tags": ["Posts"],
        "summary": "Excluir post",
        "description": "Remove permanentemente um post",
        "operationId": "excluirPost",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID do post",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          }
        ],
        "responses": {
          "204": { "description": "Post excluído com sucesso" },
          "401": { "description": "Não autorizado" },
          "403": { "description": "Permissão negada" },
          "404": { "description": "Post não encontrado" }
        }
      }
    },
    "/posts": {
      "post": {
        "tags": ["Posts"],
        "summary": "Criar novo post",
        "description": "Cria um novo post com opção de upload de arquivos",
        "operationId": "criarPost",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "post": { "$ref": "#/components/schemas/PostCreateDTO" },
                  "arquivos": {
                    "type": "array",
                    "description": "Arquivos a serem anexados ao post",
                    "items": { "type": "string", "format": "binary" }
                  }
                },
                "required": ["post"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Post criado com sucesso",
            "content": {
              "*/*": {
                "schema": { "$ref": "#/components/schemas/PostResponseDTO" }
              }
            }
          },
          "400": { "description": "Dados inválidos fornecidos" },
          "401": { "description": "Não autorizado" }
        }
      }
    },
    "/posts/{id}/salvar": {
      "post": {
        "tags": ["Posts"],
        "summary": "Salvar post",
        "description": "Salva um post, opcionalmente em uma coleção específica",
        "operationId": "salvarPost",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID do post",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          },
          {
            "name": "colecao",
            "in": "query",
            "description": "Nome da coleção para salvar o post",
            "required": false,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "201": { "description": "Post salvo com sucesso" },
          "400": { "description": "Post já salvo pelo usuário" },
          "401": { "description": "Não autorizado" },
          "404": { "description": "Post não encontrado" }
        }
      }
    },
    "/posts/{id}/curtir": {
      "post": {
        "tags": ["Posts"],
        "summary": "Curtir post",
        "description": "Adiciona uma curtida ao post",
        "operationId": "curtirPost",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID do post",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          }
        ],
        "responses": {
          "200": {
            "description": "Post curtido com sucesso",
            "content": {
              "*/*": {
                "schema": { "$ref": "#/components/schemas/PostResponseDTO" }
              }
            }
          },
          "400": { "description": "Post já curtido pelo usuário" },
          "401": { "description": "Não autorizado" },
          "404": { "description": "Post não encontrado" }
        }
      }
    },
...
    "/posts/{id}/desarquivar": {
      "patch": {
        "tags": ["Posts"],
        "summary": "Desarquivar post",
        "description": "Desarquiva um post anteriormente arquivado",
        "operationId": "desarquivarPost",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID do post",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          }
        ],
        "responses": {
          "204": { "description": "Post desarquivado com sucesso" },
          "401": { "description": "Não autorizado" },
          "403": { "description": "Permissão negada" },
          "404": { "description": "Post não encontrado" }
        }
      }
    },
    "/posts/{id}/arquivar": {
      "patch": {
        "tags": ["Posts"],
        "summary": "Arquivar post",
        "description": "Arquiva um post sem removê-lo",
        "operationId": "arquivarPost",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID do post",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          }
        ],
        "responses": {
          "204": { "description": "Post arquivado com sucesso" },
          "401": { "description": "Não autorizado" },
          "403": { "description": "Permissão negada" },
          "404": { "description": "Post não encontrado" }
        }
      }
    },
    "/posts/usuario/{usuarioId}": {
      "get": {
        "tags": ["Posts"],
        "summary": "Listar posts de um usuário",
        "description": "Retorna os posts publicados por um usuário específico",
        "operationId": "listarPostsDoUsuario",
        "parameters": [
          {
            "name": "usuarioId",
            "in": "path",
            "description": "ID do usuário",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          },
          {
            "name": "pageable",
            "in": "query",
            "description": "Parâmetros de paginação (page=0, size=10, sort=dataCriacao,desc)",
            "required": true,
            "schema": { "$ref": "#/components/schemas/Pageable" }
          }
        ],
        "responses": {
          "200": {
            "description": "Posts listados com sucesso",
            "content": {
              "*/*": { "schema": { "$ref": "#/components/schemas/Page" } }
            }
          },
          "401": { "description": "Não autorizado" },
          "404": { "description": "Usuário não encontrado" }
        }
      }
    },
    "/posts/salvos": {
      "get": {
        "tags": ["Posts"],
        "summary": "Listar posts salvos",
        "description": "Retorna os posts salvos pelo usuário logado",
        "operationId": "listarPostsSalvos",
        "parameters": [
          {
            "name": "pageable",
            "in": "query",
            "description": "Parâmetros de paginação (page=0, size=10, sort=dataCriacao,desc)",
            "required": true,
            "schema": { "$ref": "#/components/schemas/Pageable" }
          }
        ],
        "responses": {
          "200": {
            "description": "Posts salvos listados com sucesso",
            "content": {
              "*/*": { "schema": { "$ref": "#/components/schemas/Page" } }
            }
          },
          "401": { "description": "Não autorizado" }
        }
      }
    },
    "/posts/salvos/colecao/{colecao}": {
      "get": {
        "tags": ["Posts"],
        "summary": "Listar posts salvos por coleção",
        "description": "Retorna posts salvos pelo usuário em uma coleção específica",
        "operationId": "listarPostsSalvosPorColecao",
        "parameters": [
          {
            "name": "colecao",
            "in": "path",
            "description": "Nome da coleção",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "pageable",
            "in": "query",
            "description": "Parâmetros de paginação (page=0, size=10, sort=dataCriacao,desc)",
            "required": true,
            "schema": { "$ref": "#/components/schemas/Pageable" }
          }
        ],
        "responses": {
          "200": {
            "description": "Posts da coleção listados com sucesso",
            "content": {
              "*/*": { "schema": { "$ref": "#/components/schemas/Page" } }
            }
          },
          "401": { "description": "Não autorizado" }
        }
      }
    },
    "/posts/populares": {
      "get": {
        "tags": ["Posts"],
        "summary": "Listar posts populares",
        "description": "Retorna os posts mais populares da plataforma",
        "operationId": "listarPostsPopulares",
        "parameters": [
          {
            "name": "pageable",
            "in": "query",
            "description": "Parâmetros de paginação (page=0, size=10, sort=dataCriacao,desc)",
            "required": true,
            "schema": { "$ref": "#/components/schemas/Pageable" }
          }
        ],
        "responses": {
          "200": {
            "description": "Posts populares listados com sucesso",
            "content": {
              "*/*": { "schema": { "$ref": "#/components/schemas/Page" } }
            }
          },
          "401": { "description": "Não autorizado" }
        }
      }
    },
    "/posts/localizacao": {
      "get": {
        "tags": ["Posts"],
        "summary": "Listar posts por localização",
        "description": "Retorna posts associados a uma localização específica",
        "operationId": "listarPostsPorLocalizacao",
        "parameters": [
          {
            "name": "local",
            "in": "query",
            "description": "Nome da localização",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "pageable",
            "in": "query",
            "description": "Parâmetros de paginação (page=0, size=10, sort=dataCriacao,desc)",
            "required": true,
            "schema": { "$ref": "#/components/schemas/Pageable" }
          }
        ],
        "responses": {
          "200": {
            "description": "Posts listados com sucesso",
            "content": {
              "*/*": { "schema": { "$ref": "#/components/schemas/Page" } }
            }
          },
          "401": { "description": "Não autorizado" }
        }
      }
    },
    "/posts/hashtag/{nome}": {
      "get": {
        "tags": ["Posts"],
        "summary": "Listar posts por hashtag",
        "description": "Retorna posts que contêm uma hashtag específica",
        "operationId": "listarPostsPorHashtag",
        "parameters": [
          {
            "name": "nome",
            "in": "path",
            "description": "Nome da hashtag",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "pageable",
            "in": "query",
            "description": "Parâmetros de paginação (page=0, size=10, sort=dataCriacao,desc)",
            "required": true,
            "schema": { "$ref": "#/components/schemas/Pageable" }
          }
        ],
        "responses": {
          "200": {
            "description": "Posts listados com sucesso",
            "content": {
              "*/*": { "schema": { "$ref": "#/components/schemas/Page" } }
            }
          },
          "401": { "description": "Não autorizado" }
        }
      }
    },
    "/posts/feed": {
      "get": {
        "tags": ["Posts"],
        "summary": "Listar feed do usuário",
        "description": "Retorna o feed personalizado do usuário logado",
        "operationId": "listarFeed",
        "parameters": [
          {
            "name": "pageable",
            "in": "query",
            "description": "Parâmetros de paginação (page=0, size=10, sort=dataCriacao,desc)",
            "required": true,
            "schema": { "$ref": "#/components/schemas/Pageable" }
          }
        ],
        "responses": {
          "200": {
            "description": "Feed listado com sucesso",
            "content": {
              "*/*": { "schema": { "$ref": "#/components/schemas/Page" } }
            }
          },
          "401": { "description": "Não autorizado" }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "PostUpdateDTO": {
        "type": "object",
        "description": "Dados para atualização do post",
        "properties": {
          "legenda": { "type": "string", "maxLength": 2200, "minLength": 0 },
          "localizacao": { "type": "string" },
          "visibilidade": {
            "type": "string",
            "enum": ["PUBLICO", "PRIVADO", "SOMENTE_SEGUIDORES"]
          },
          "hashtags": { "type": "array", "items": { "type": "string" } },
          "usuariosMarcados": {
            "type": "array",
            "items": { "type": "integer", "format": "int64" }
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "Insira um token JWT válido aqui. Expira em 24h.",
        "name": "bearerAuth",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "refreshToken": {
        "type": "http",
        "description": "Insira um refresh token válido para obter novos tokens. Validade de 7 dias.",
        "name": "refreshToken",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}